+++ lib/definitions/src/api/apierror.d.ts
@@ -1,6 +1,0 @@
-export declare class ApiError {
-    domain: string;
-    reason: string;
-    message: string;
-    constructor(reason: string, message: string);
-}