+++ lib/definitions/src/test/model/user.d.ts
@@ -1,9 +1,0 @@
-import { Model } from '../../base/model';
-export declare class User extends Model {
-    uid: number;
-    username: string;
-    displayName: string;
-    meta: any;
-    createdAt: Date;
-    updatedAt: number;
-}