+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.1.9 (2016.12.21.)
+
+* **SuccessResponse 添加toJSON**
+* **export 了 ErrorResponse**
+
 # 2.1.8 (2016.12.20.)
 
 * **开始发送邮件给开发者:** back@gagogroup.com。